refactor(nvim): insert "ethersync" to each module
not to pollute lua package.loaded namespaces with general terms.
