+++ nvim-plugin/lua/ethersync/logging.lua
